+++ app/src/main/java/woowacourse/shopping/view/detail/ProductDetailFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view
+package woowacourse.shopping.view.detail
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
