+++ app/src/main/res/layout/fragment_product_detail.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.ProductDetailFragment">
+        tools:context=".view.detail.ProductDetailFragment">
 
         <ImageView
             android:id="@+id/iv_detail"
