+++ app/src/main/res/layout/fragment_product_list.xml
@@ -15,5 +15,5 @@
         android:layout_height="match_parent"
         android:padding="14dp"
         app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        tools:context=".view.ProductListFragment" />
+        tools:context=".view.products.ProductListFragment" />
 </layout>
