+++ app/src/main/res/layout/fragment_shopping_cart.xml
@@ -17,7 +17,7 @@
             android:layout_height="match_parent"
             android:orientation="vertical"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            tools:context=".view.ShoppingCartFragment"
+            tools:context=".view.cart.ShoppingCartFragment"
             tools:itemCount="3"
             tools:layout_editor_absoluteX="19dp"
             tools:layout_editor_absoluteY="0dp"
